requirements.txt
python-binance==1.0.32
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
torch==2.3.0 --index-url https://download.pytorch.org/whl/cpu
python-dotenv==1.0.1
matplotlib==3.8.4
rich==13.7.1
requests==2.31.0
optuna==3.6.1
flask==3.0.3
websocket-client==1.7.0
